+++ R/imports.R
@@ -8,4 +8,5 @@
 #' @importFrom Rcpp sourceCpp
 #' @import graphics
 #' @import stats
+#' @importFrom dplyr group_by mutate %>% group_map
 NULL